Updating tests for networkservice
