+++ Tests/NetworkServiceTests/MockURLProtocol.swift
@@ -1,0 +1,36 @@
+import Foundation
+
+class MockURLProtocol: URLProtocol {
+    
+    static var requestHandler: ((URLRequest) throws -> (HTTPURLResponse, Data?))?
+    
+    override class func canInit(with request: URLRequest) -> Bool {
+        return true
+    }
+    
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+        return request
+    }
+    
+    override func startLoading() {
+        guard let handler = MockURLProtocol.requestHandler else {
+            fatalError("MockURLProtocol.requestHandler is not set.")
+        }
+        
+        do {
+            let (response, data) = try handler(request)
+            
+            if let data = data {
+                client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+                client?.urlProtocol(self, didLoad: data)
+            }
+            
+            client?.urlProtocolDidFinishLoading(self)
+        } catch {
+            client?.urlProtocol(self, didFailWithError: error)
+        }
+    }
+    
+    override func stopLoading() {
+    }
+}
